make worker code a sub package
