+++ src/index.ts
@@ -1,12 +1,2 @@
-// client
-export { subscribeToPush, unsubscribeFromPush, getPushSubscriptionStatus } from './client/subscription';
-export type { PushRoute } from './client/subscription';
-
-// push 
-export { Push } from './push/strategy'
-
-// types
-export type { NotificationObject, VapidDetails, PushHandlerEnv, PushPlugin, PushType } from './types';
-
-export { sendNotifications } from './server/server';
-export type { sendNotificationParams } from './server/server';
+export * from './client';
+export * from './server';
