+++ src/server/send-notifications.ts
@@ -1,4 +1,17 @@
-import { NotificationObject, VapidDetails } from '../types';
+export type VapidDetails = {
+  publicKey: string;
+  privateKey: string;
+  subject: string;
+};
+
+export interface NotificationObjectOptions extends NotificationOptions {
+  body: string;
+}
+
+export type NotificationObject = {
+  title: string;
+  options: NotificationObjectOptions;
+};
 
 export type sendNotificationParams = {
   subscriptions: PushSubscription[];
