+++ src/worker/push/types.ts
@@ -1,22 +1,3 @@
-/**
- * @fileoverview Types used across the package. To be re-exported from `index.ts`
- */
-
-export type VapidDetails = {
-  publicKey: string;
-  privateKey: string;
-  subject: string;
-};
-
-export interface NotificationObjectOptions extends NotificationOptions {
-  body: string;
-}
-
-export type NotificationObject = {
-  title: string;
-  options: NotificationObjectOptions;
-};
-
 export interface PushHandlerEnv {
   event: PushEvent | NotificationEvent | ErrorEvent;
   state?: Record<string, any>;
